+++ changelog.md
@@ -1,7 +1,7 @@
 Changelog for the [`directory`][1] package
 ==========================================
 
-## 1.2.3.0 (June 2015)
+## 1.2.3.0 (July 2015)
 
   * Add support for XDG Base Directory Specification
     ([#6](https://github.com/haskell/directory/issues/6))